// src/frontend/ui/generated/PlanList.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PlanListWorkspace } from "./PlanListWorkspace";
import type { PlanPb } from "./PlanPb";

/**
 * `plan-list` 的回包。
 */
export type PlanList = { pb: PlanPb, workspaces: Array<PlanListWorkspace>, };